Lightmap baked for Nørreport scene. See comments...
There is a catch to the baked lightmap. To avoid having a model for each tracker, and having their bake affect each other, only one model is in the scene with the bake. Unfortunately, duplicating the object does NOT copy the lightmap. 
It should be manageable though, if we only use one marker-setup - Just make sure the object is parented to the right tracker.